+++ derender/test/derender/node_test.rb
@@ -5,6 +5,26 @@ require_relative "../test_helper"
 module Sevgi
   module Derender
     class NodeTest < Minitest::Test
+      def test_node_owns_immutable_public_state
+        node = Derender.decompile("<svg xmlns=\"urn:test\" id=\"root\"><g id=\"child\"/></svg>")
+
+        [node, node.attributes, node.namespaces, node.meta, node.children, node.content, node.name].each do |value|
+          assert_predicate(value, :frozen?)
+        end
+
+        [*node.attributes, *node.namespaces, *node.meta].flatten.each { assert_predicate(it, :frozen?) }
+        assert_predicate(node.children.first, :frozen?)
+      end
+
+      def test_node_hides_parser_and_strategy_plumbing
+        node = Derender.decompile("<svg/>")
+
+        assert_raises(NameError) { Derender::Document }
+        refute_respond_to(Node, :new)
+        %i[node type pres decompile attributes!].each { refute_respond_to(node, it) }
+        assert_empty(Derender.public_instance_methods(false))
+      end
+
       def test_dispatch_respects_namespace_and_conversion_root
         cases = {
           "<style>.a { fill: red; }</style>" => :CSS,
@@ -20,17 +40,17 @@ module Sevgi
         }
 
         cases.each do |xml, type|
-          assert_equal(type, Derender::Document.new(xml).decompile.type, xml)
+          assert_equal(type, Document.new(xml).decompile.send(:type), xml)
         end
 
         xml = "<svg xmlns=\"http://www.w3.org/2000/svg\"><svg id=\"nested\"/></svg>"
-        root = Derender::Document.new(xml).decompile
+        root = Document.new(xml).decompile
 
-        assert_equal(:Any, root.children.first.type)
-        assert_equal(:Root, Derender::Document.new(xml).decompile("nested").type)
+        assert_equal(:Any, root.children.first.send(:type))
+        assert_equal(:Root, Document.new(xml).decompile("nested").send(:type))
 
-        foreign_style = Derender::Document.new("<style xmlns=\"urn:foreign\">raw</style>").decompile.derender
-        foreign_svg = Derender::Document.new("<svg xmlns=\"urn:foreign\"/>").decompile.derender
+        foreign_style = Document.new("<style xmlns=\"urn:foreign\">raw</style>").decompile.derender
+        foreign_svg = Document.new("<svg xmlns=\"urn:foreign\"/>").decompile.derender
 
         assert_match(/\AElement\(:"style",/, foreign_style)
         assert_match(/\AElement\(:"svg",/, foreign_svg)
@@ -45,7 +65,7 @@ module Sevgi
         SVG
           .chomp
 
-        actual = Derender::Document.new(svg).decompile("xxx").attributes
+        actual = Document.new(svg).decompile("xxx").attributes
         expected = {"id" => "xxx", "_:foo" => "fff", "_:bar" => "bbb"}
 
         assert_equal(expected, actual)
@@ -60,7 +80,7 @@ module Sevgi
         SVG
           .chomp
 
-        actual = Derender::Document.new(svg).decompile("xxx")._
+        actual = Document.new(svg).decompile("xxx")._
         expected = {"foo" => "fff", "bar" => "bbb"}
 
         assert_equal(expected, actual)
@@ -75,7 +95,7 @@ module Sevgi
         SVG
           .chomp
 
-        actual = Derender::Document.new(svg).decompile.find("line1").derender
+        actual = Document.new(svg).decompile.find("line1").derender
 
         expected = <<~SEVGI
           line id: "line1", length: 10.0
@@ -90,7 +110,7 @@ module Sevgi
         SVG
           .chomp
 
-        assert_nil(Derender::Document.new(svg).decompile.find("line1"))
+        assert_nil(Document.new(svg).decompile.find("line1"))
       end
 
       def test_find_returns_root_node
@@ -102,7 +122,7 @@ module Sevgi
         SVG
           .chomp
 
-        actual = Derender::Document.new(svg).decompile.find("xxx").derender
+        actual = Document.new(svg).decompile.find("xxx").derender
 
         expected = <<~SEVGI
           g id: "xxx", "_:foo": "fff", "_:bar": "bbb" do
@@ -123,7 +143,7 @@ module Sevgi
         SVG
           .chomp
 
-        actual = Derender::Document.new(svg).decompile.find("20.0", by: "length").derender
+        actual = Document.new(svg).decompile.find("20.0", by: "length").derender
 
         expected = <<~SEVGI
           line id: "line2", length: 20.0